+++ gribpcraster/application/domain/GribGridDetails.py
@@ -4,17 +4,7 @@ from util.logger.Logger import Logger
 import gribpcraster.application.ExecutionContext as ex
 
 #TODO AVOID USAGE OF grib_get. Build id and infos with extractWorkingKeys once!
-def _getShape(gid, grid_type_, numValues):
-    nx = ny = 0
-    if grid_type_ in ('regular_gg','regular_ll','rotated_ll','rotated_gg'):  # regular lat/lon grid
-        nx = GRIB.grib_get(gid, 'Ni')
-        ny = GRIB.grib_get(gid, 'Nj')
-    elif grid_type_ in ('reduced_gg', 'reduced_ll'):  # reduced global gaussian grid
-        ny = GRIB.grib_get(gid, 'Nj')
-        nx = numValues/ny
 
-    shape = (nx, ny)
-    return shape
 
 def _buildId(gid, grid_type):
 
@@ -36,6 +26,18 @@ class GribGridDetails(object):
     def _log(self, message, level='DEBUG'):
         self._logger.log(message, level)
 
+    def _getShape(self):
+        nx = ny = 0
+        if self._grid_type in ('regular_gg','regular_ll','rotated_ll','rotated_gg'):  # regular lat/lon grid
+            nx = self._geo_keys.get('Ni')
+            ny = self._geo_keys.get('Nj')
+        elif self._grid_type in ('reduced_gg', 'reduced_ll'):  # reduced global gaussian grid
+            ny = self._geo_keys.get('Nj')
+            nx = self._geo_keys.get('numberOfValues')/ny
+
+        shape = (nx, ny)
+        return shape
+
     def __init__(self, gid):
 
         #Managed grid types:
@@ -45,16 +47,16 @@ class GribGridDetails(object):
         self._lats = self._longs = None
         self._logger = Logger('Messages', loggingLevel=ex.global_logger_level)
         self._gid = gid
-        self._grid_type = GRIB.grib_get(gid, 'gridType')
-        self._geo_keys = self._extractWorkingKeys(gid)
-        self._missing_value = GRIB.grib_get(gid, 'missingValue')
-        self._shape = _getShape(gid, self._grid_type, GRIB.grib_get(gid, 'numberOfValues'))
+        self._geo_keys = self._extract_info_keys(gid)
+        self._grid_type = self._geo_keys.get('gridType')
+        self._points_meridian = self._geo_keys.get('Nj')
+        self._missing_value = self._geo_keys.get('missingValue')
+        self._shape = self._getShape()
         #lazy computation
         self._lats = None
         self._longs = None
 
         self._grid_id = _buildId(gid,self._grid_type)
-        self._points_meridian = GRIB.grib_get(gid,'Nj')
 
         self._grid_details_2nd = None
         self._change_resolution_step = None
@@ -99,59 +101,26 @@ class GribGridDetails(object):
             self._lats, self._longs = self._computeLatLongs(self._gid)
         return self._lats, self._longs
 
-    def _extractWorkingKeys(self, gid):
+    def _extract_info_keys(self, gid):
 
         working_keys = {}
+        if GRIB.grib_is_defined(gid, 'gridType'):
+            working_keys['gridType'] = GRIB.grib_get_string(gid, 'gridType')
 
         if GRIB.grib_is_defined(gid, 'radius'):
-            working_keys['radius'] = GRIB.grib_get(gid, 'radius')
-        # working_keys['shapeOfTheEarth'] = GRIB.grib_get(gid, 'shapeOfTheEarth')
-        # if GRIB.grib_is_defined(gid, 'scaleFactorOfMajorAxisOfOblateSpheroidEarth'):
-        #     working_keys['scaleFactorOfMajorAxisOfOblateSpheroidEarth'] = GRIB.grib_get(gid,
-        #                                                                                 'scaleFactorOfMajorAxisOfOblateSpheroidEarth')
+            working_keys['radius'] = GRIB.grib_get_double(gid, 'radius')
+
+        if GRIB.grib_is_defined(gid, 'numberOfValues'):
+            working_keys['numberOfValues'] = GRIB.grib_get_long(gid, 'numberOfValues')
+
+        if GRIB.grib_is_defined(gid, 'Ni'):
+            working_keys['Ni'] = GRIB.grib_get_long(gid, 'Ni')
+
+        if GRIB.grib_is_defined(gid, 'Nj'):
+            working_keys['Nj'] = GRIB.grib_get_long(gid, 'Nj')
+
         if GRIB.grib_is_defined(gid, 'missingValue'):
             working_keys['missingValue'] = GRIB.grib_get_double(gid, 'missingValue')
-
-        # scaleFactorOfRadiusOfSphericalEarth
-        # if GRIB.grib_is_defined(gid, 'scaleFactorOfRadiusOfSphericalEarth'):
-        #     working_keys['scaleFactorOfRadiusOfSphericalEarth'] = GRIB.grib_get_double(gid,
-        #                                                                                'scaleFactorOfRadiusOfSphericalEarth')
-        #
-        # # scaledValueOfMajorAxisOfOblateSpheroidEarth
-        # if GRIB.grib_is_defined(gid, 'scaledValueOfMajorAxisOfOblateSpheroidEarth'):
-        #     working_keys['scaledValueOfMajorAxisOfOblateSpheroidEarth'] = GRIB.grib_get_double(gid,
-        #                                                                                        'scaledValueOfMajorAxisOfOblateSpheroidEarth')
-        #     # scaleFactorOfMinorAxisOfOblateSpheroidEarth
-        # if GRIB.grib_is_defined(gid, 'scaleFactorOfMinorAxisOfOblateSpheroidEarth'):
-        #     working_keys['scaleFactorOfMinorAxisOfOblateSpheroidEarth'] = GRIB.grib_get_double(gid,
-        #                                                                                        'scaleFactorOfMinorAxisOfOblateSpheroidEarth')
-        #     # scaledValueOfEarthMajorAxis
-        # if GRIB.grib_is_defined(gid, 'scaledValueOfEarthMajorAxis'):
-        #     working_keys['scaledValueOfEarthMajorAxis'] = GRIB.grib_get_double(gid, 'scaledValueOfEarthMajorAxis')
-        #     # scaledValueOfEarthMinorAxis
-        # if GRIB.grib_is_defined(gid, 'scaledValueOfEarthMinorAxis'):
-        #     working_keys['scaledValueOfEarthMinorAxis'] = GRIB.grib_get_double(gid, 'scaledValueOfEarthMinorAxis')
-        #
-        # # scaledValueOfRadiusOfSphericalEarth
-        # if GRIB.grib_is_defined(gid, 'scaledValueOfRadiusOfSphericalEarth'):
-        #     working_keys['scaledValueOfRadiusOfSphericalEarth'] = GRIB.grib_get_double(gid,
-        #                                                                                'scaledValueOfRadiusOfSphericalEarth')
-        #     # standardParallel
-        # if GRIB.grib_is_defined(gid, 'standardParallel'):
-        #     working_keys['standardParallel'] = GRIB.grib_get_double(gid, 'standardParallel')
-        #     # centralLongitude
-        # if GRIB.grib_is_defined(gid, 'centralLongitude'):
-        #     working_keys['centralLongitude'] = GRIB.grib_get_double(gid, 'centralLongitude')
-        #     # angleOfRotationInDegrees
-        # if GRIB.grib_is_defined(gid, 'angleOfRotationInDegrees'):
-        #     working_keys['angleOfRotationInDegrees'] = GRIB.grib_get_double(gid, 'angleOfRotationInDegrees')
-        #     # latitudeOfSouthernPoleInDegrees
-        # if GRIB.grib_is_defined(gid, 'latitudeOfSouthernPoleInDegrees'):
-        #     working_keys['latitudeOfSouthernPoleInDegrees'] = GRIB.grib_get_double(gid,
-        #                                                                            'latitudeOfSouthernPoleInDegrees')
-        #     # longitudeOfSouthernPoleInDegrees
-        # if GRIB.grib_is_defined(gid, 'longitudeOfSouthernPoleInDegrees'):
-        #     working_keys['longitudeOfSouthernPoleInDegrees'] = GRIB.grib_get_double(gid, 'longitudeOfSouthernPoleInDegrees')
 
         return working_keys